temporary- make inmem feature store look like master
